Drop Domain.MongoDB project and move configuration to Data project
