+++ Defra.TradeImportsDataApi.sln
@@ -19,9 +19,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Data", "src\Data\Data.cspro
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Domain", "src\Domain\Domain.csproj", "{9457F847-9127-466B-ADDA-99DE6F5FA38A}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Domain.MongoDB.Tests", "tests\Domain.MongoDB.Tests\Domain.MongoDB.Tests.csproj", "{CDEC7EC0-D0DA-474A-A327-5D27CDFEAD58}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Domain.MongoDB", "src\Domain.MongoDB\Domain.MongoDB.csproj", "{0C4ED0EE-75A3-40CC-AE6D-556215D90BA4}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Data.Tests", "tests\Data.Tests\Data.Tests.csproj", "{CDEC7EC0-D0DA-474A-A327-5D27CDFEAD58}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Api.Client", "src\Api.Client\Api.Client.csproj", "{30A2F6EC-0A32-4685-97A1-9192863ADE51}"
 EndProject
@@ -59,10 +57,6 @@ Global
 		{CDEC7EC0-D0DA-474A-A327-5D27CDFEAD58}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{CDEC7EC0-D0DA-474A-A327-5D27CDFEAD58}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{CDEC7EC0-D0DA-474A-A327-5D27CDFEAD58}.Release|Any CPU.Build.0 = Release|Any CPU
-		{0C4ED0EE-75A3-40CC-AE6D-556215D90BA4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0C4ED0EE-75A3-40CC-AE6D-556215D90BA4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0C4ED0EE-75A3-40CC-AE6D-556215D90BA4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0C4ED0EE-75A3-40CC-AE6D-556215D90BA4}.Release|Any CPU.Build.0 = Release|Any CPU
 		{30A2F6EC-0A32-4685-97A1-9192863ADE51}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{30A2F6EC-0A32-4685-97A1-9192863ADE51}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{30A2F6EC-0A32-4685-97A1-9192863ADE51}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -79,7 +73,6 @@ Global
 		{28337F7B-6FC1-41B0-8D55-2707408923E4} = {30577215-A7B4-4DBB-9F6B-38D2AA88E7E2}
 		{9457F847-9127-466B-ADDA-99DE6F5FA38A} = {30577215-A7B4-4DBB-9F6B-38D2AA88E7E2}
 		{CDEC7EC0-D0DA-474A-A327-5D27CDFEAD58} = {92349C3B-0C37-4727-ADAC-95748294D315}
-		{0C4ED0EE-75A3-40CC-AE6D-556215D90BA4} = {30577215-A7B4-4DBB-9F6B-38D2AA88E7E2}
 		{30A2F6EC-0A32-4685-97A1-9192863ADE51} = {30577215-A7B4-4DBB-9F6B-38D2AA88E7E2}
 	EndGlobalSection
 EndGlobal
