+++ src/Data/Extensions/ServiceCollectionExtensions.cs
@@ -1,5 +1,5 @@
+using Defra.TradeImportsDataApi.Data.Configuration;
 using Defra.TradeImportsDataApi.Data.Mongo;
-using Defra.TradeImportsDataApi.Domain.MongoDB;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
@@ -40,7 +40,7 @@ public static class ServiceCollectionExtensions
             };
 
             ConventionRegistry.Register(nameof(conventionPack), conventionPack, _ => true);
-            ClassMapConfiguration.Register();
+            DomainClassMapConfiguration.Register();
 
             return client.GetDatabase(options?.Value.DatabaseName);
         });
